add persistence to gruendung-context
