+++ apps/frontend/src/App.tsx
@@ -10,47 +10,48 @@ import GetEntryPage from "./pages/GetDienstleisterPage/GetDienstleisterPage";
 import GruendungContextProvider from "contexts/Gruendung/Gruendung";
 import Impressum from "./pages/Impressum/Impressum";
 import Datenschutzerklaerung from "./pages/Datenschutzerklaerung/Datenschutzerklaerung";
+import Gruendung_TaskId from "./pages/Gruendung/$task";
 
 function App() {
 	return (
 		<AuthProvider>
-			<Router>
-				<Header />
-				<Switch>
-					<Route path='/' exact>
-						<Index />
-					</Route>
-					<Route path='/gruendung' exact>
-						<Index />
-					</Route>
-					<Route path='/gruendung/:task'>
-						<GruendungContextProvider>
+			<GruendungContextProvider>
+				<Router>
+					<Header />
+					<Switch>
+						<Route path='/' exact>
+							<Index />
+						</Route>
+						<Route path='/gruendung' exact>
 							<Gruendung />
-						</GruendungContextProvider>
-					</Route>
-					<Route path='/dienstleister'>
-						<Index />
-					</Route>
-					<Route path='/dienstleister/search'>
-						<Search />
-					</Route>
-					<Route path='/dienstleister/create'>
-						<CreateEntryPage />
-					</Route>
-					<Route path='/dienstleister/get'>
-						<GetEntryPage />
-					</Route>
-					<Route path='/demoAuth'>
-						<DemoAuth />
-					</Route>
-					<Route path='/impressum'>
-						<Impressum />
-					</Route>
-					<Route path='/datenschutz'>
-						<Datenschutzerklaerung />
-					</Route>
-				</Switch>
-			</Router>
+						</Route>
+						<Route path='/gruendung/:task'>
+							<Gruendung_TaskId />
+						</Route>
+						<Route path='/dienstleister'>
+							<Index />
+						</Route>
+						<Route path='/dienstleister/search'>
+							<Search />
+						</Route>
+						<Route path='/dienstleister/create'>
+							<CreateEntryPage />
+						</Route>
+						<Route path='/dienstleister/get'>
+							<GetEntryPage />
+						</Route>
+						<Route path='/demoAuth'>
+							<DemoAuth />
+						</Route>
+						<Route path='/impressum'>
+							<Impressum />
+						</Route>
+						<Route path='/datenschutz'>
+							<Datenschutzerklaerung />
+						</Route>
+					</Switch>
+				</Router>
+			</GruendungContextProvider>
 		</AuthProvider>
 	);
 }
