+++ apps/frontend/src/components/TaskList/Nodes/Nodes.tsx
@@ -1,5 +1,4 @@
 import {
-	DecisionNode,
 	ProcessedTaskNode,
 	useGruendungContext,
 } from "contexts/Gruendung/Gruendung";
